+++ ChangeLog.md
@@ -4,14 +4,13 @@ All notable changes to stumpless will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
 For a detailed look at the project's future, including planned features and bug
 fixes, check out the
 [roadmap](https://github.com/goatshriek/stumpless/blob/master/docs/roadmap.md).
-### Added 
- - @since format check enforcement in CI pipeline.
 
 ## [2.2.0] - 2023-08-08
+### Added 
+ - @since format check enforcement in CI pipeline.
 ### Fixed
  - Deadlock potential in `stumpless_set_entry_hostname` and
    `stumpless_set_entry_procid` when validation fails.