+++ ChangeLog.md
@@ -8,9 +8,11 @@ For a detailed look at the project's future, including planned features and bug
 fixes, check out the
 [roadmap](https://github.com/goatshriek/stumpless/blob/master/docs/roadmap.md).
 
-## [2.2.0] - 2023-08-08
+
+## [2.2.0] - 2023-10-14
 ### Added 
  - @since format check enforcement in CI pipeline.
+ - `single-file` target for rollup `.c` and `.h` files.
 ### Fixed
  - Deadlock potential in `stumpless_set_entry_hostname` and
    `stumpless_set_entry_procid` when validation fails.